+++ src/Annotation/Inject.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Annotation;
 
-/**
- * @Annotation
- */
+#[Attribute(Attribute::TARGET_PROPERTY)]
 final class Inject
 {
 }