+++ tests/Fixture/DependencyGraph/Attribute/AttributeD.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WoohooLabs\Zen\Tests\Fixture\DependencyGraph\Attribute;
+
+class AttributeD
+{
+}