+++ tests/Fixture/DependencyGraph/Exception/ExceptionA.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Tests\Fixture\DependencyGraph\Exception;
 
+use WoohooLabs\Zen\Annotation\Inject;
+
 class ExceptionA
 {
-    /**
-     * @Inject
-     */
+    #[Inject]
     private $a;
 }