+++ tests/Fixture/DependencyGraph/Exception/ExceptionF.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Tests\Fixture\DependencyGraph\Exception;
 
+use WoohooLabs\Zen\Annotation\Inject;
+
 class ExceptionF
 {
-    /**
-     * @Inject
-     * @var ExceptionA
-     */
+    #[Inject]
+    /** @var ExceptionA */
     protected static $a;
 }