+++ tests/Fixture/DependencyGraph/Mixed/MixedC.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Tests\Fixture\DependencyGraph\Mixed;
 
+use WoohooLabs\Zen\Annotation\Inject;
+
 class MixedC
 {
-    /**
-     * @Inject
-     * @var MixedB
-     */
+    #[Inject]
+    /** @var MixedB */
     private $b;
 }